requirements.txt
flask
snowflake-connector-python
openai==0.28
python-dotenv
streamlit